src/requirements.txt
APScheduler==3.11.0
ccxt==3.0.76
discord.py==2.2.2
firebase_admin==6.6.0
google_search_results==2.4.2
langchain_core==0.3.33
langchain_openai==0.3.3
langgraph==0.2.69
python-dotenv==1.0.1
pytz==2024.2
Requests==2.32.3
pydantic==2.10.6
